requirements.txt
flask
joblib
pandas
flask_bootstrap
werkzeug
Pillow
numpy
opencv-python
pandas
matplotlib
sklearn
seaborn
utils
